cleanups

+++ src/encode.h
@@ -24,8 +24,8 @@
 
 typedef int ReadSampleFunction(void *data, int16_t *sample);
 
-int32_t acm_encode(ReadSampleFunction *read, void *data, FILE *out, unsigned channels,
-		   unsigned sample_rate, float volume, int levels, int samples_per_subband,
+int32_t acm_encode(ReadSampleFunction *read, void *data, FILE *out, uint16_t channels,
+		   uint32_t sample_rate, float volume, int levels, int samples_per_subband,
 		   float comp_ratio, int wavc);
 
 #endif
